+++ src/open_links.js
@@ -1,5 +1,6 @@
 const vscode = require('vscode');
 const opn = require('opn');
+const gitService = require('./git_service');
 const gitLabService = require('./gitlab_service');
 
 const { instanceUrl } = vscode.workspace.getConfiguration('gitlab');
@@ -22,5 +23,36 @@ async function showMergeRequests() {
   await openLink(`${instanceUrl}/dashboard/merge_requests?assignee_id=$userId`);
 };
 
+async function openActiveFileOnWeb() {
+  const editor = vscode.window.activeTextEditor;
+
+  if (editor) {
+    const currentProject = await gitLabService.fetchCurrentProject();
+
+    if (currentProject) {
+      const branchName = await gitService.fetchBranchName();
+      const filePath = editor.document.uri.path.replace(vscode.workspace.rootPath, '');
+      let fileUrl = `${currentProject.web_url}/blob/${branchName}/${filePath}`;
+      let anchor = '';
+
+      if (editor.selection) {
+        const { start, end } = editor.selection;
+        anchor = `#L${start.line + 1}`;
+
+        if (end.line > start.line) {
+          anchor += `-${end.line + 1}`;
+        }
+      }
+
+      opn(`${fileUrl}${anchor}`);
+    } else {
+      vscode.window.showInformationMessage('GitLab Workflow: Failed to open file on web. No GitLab project.');
+    }
+  } else {
+    vscode.window.showInformationMessage('GitLab Workflow: No open file.');
+  }
+}
+
 exports.showIssues = showIssues;
 exports.showMergeRequests = showMergeRequests;
+exports.openActiveFileOnWeb = openActiveFileOnWeb;